Kisebb finomítások
- A for ciklus-ban nem hívjuk meg mindig az unbindet csak a for ciklus után
- setDefaultColor uint8 típusú számot kér, hogy egyértelmű legyen hogy [0;255] intervallumban kéri a színeket

+++ include/core/rendering/renderer.h
@@ -55,7 +55,7 @@ public:
     ~Renderer();
 
     void setShadingMode(Shaders::SHADINGMODE shadingMode);
-    void setDefaultColor(float r, float g, float b);
+    void setDefaultColor(uint8_t r, uint8_t g, uint8_t b);
 
     void setImageDimensions(int imageW, int imageH);
     void addNewShader(Shaders::SHADINGMODE mode, std::unique_ptr<Shaders::Shader> shader);
